consolidate botuils and utils
